feat: use sass-embedded
