+++ .env.local-tutor
@@ -1,6 +1,6 @@
 API_URL='http://edly.io:8888'
 ELEVENTY_ENV='local:tutor'
-NODE_ENV='local'
+NODE_ENV='production'
 ROOT_DOMAIN='edly.io'
 ROOT_URL='http://edly.io:8888'
 STATIC_ASSETS_URL='http://edly.io:8888'
