+++ package.json
@@ -16,8 +16,8 @@
     "dev:tutor": "dotenvx run -f .env.dev-tutor -- npx netlify dev",
     "local:tutor": "dotenvx run -f .env.local-tutor -- npx netlify dev",
     "start": "dotenvx run -- npm run clean && npx @11ty/eleventy --serve --watch",
-    "build:dev": "NODE_ENV=development ELEVENTY_ENV=development dotenvx run -f .env.development -- npm run clean && npx @11ty/eleventy",
-    "build": "NODE_ENV=production ELEVENTY_ENV=production dotenvx run -- npm run clean && npx @11ty/eleventy"
+    "build:dev": "export NODE_ENV=development; export ELEVENTY_ENV=development; dotenvx run -f .env.development -- npm run clean && npx @11ty/eleventy",
+    "build": "export NODE_ENV=production; export ELEVENTY_ENV=production; dotenvx run -- npm run clean && npx @11ty/eleventy"
   },
   "keywords": [
     "Eleventy"
@@ -45,7 +45,7 @@
     "npm-run-all": "^4.1.5",
     "rimraf": "^6.0.1",
     "rss-parser": "^3.13.0",
-    "sass": "^1.87.0"
+    "sass-embedded": "^1.87.0"
   },
   "overrides": {
     "esbuild": "0.25.3"
